requirements.txt
discord.py~=1.6.0
requests~=2.25.1
numpy~=1.20.1
pytz~=2021.1
aiohttp~=3.7.4
pymongo[tls,srv]~=3.11.3
gspread~=3.7.0
Pillow~=8.1.1
python_dateutil~=2.8.1
schedule~=1.0.0